requirements.txt
Django~=3.1.1
djangorestframework~=3.11.1
psycopg2-binary>=2.8
social-auth-app-django
Pillow~=7.2.0